Rewrite Status with TS + some fixes
