+++ src/components/__input/Input.tsx
@@ -1,7 +1,7 @@
 import React from "react";
+import { IInput } from "../types";
 import "./--orange.css";
 import "./--blue.css";
-
 
 /**
  * This component represents a Button
@@ -58,16 +58,3 @@ export const Input = React.forwardRef<HTMLInputElement, IInput>(
         );
     }
 );
-
-type MainStyle = "orange" | "blue";
-
-interface IInput {
-    name: string;
-    style: MainStyle;
-    classNames: string;
-    classNamesBox: string;
-    onChange: (val: React.ChangeEvent) => void;
-    children: React.ReactNode;
-    type: string;
-    inputName: string;
-}
