+++ src/components/types.ts
@@ -1,6 +1,24 @@
-export type ComponentStyle = "orange" | "blue";
-export type ComponentSize = "small" | "medium" | "big";
+type ComponentStyle = "orange" | "blue";
+type ComponentSize = "small" | "medium" | "big";
 
+export interface IButton {
+    onClick?: () => void;
+    style?: ComponentStyle;
+    classNames?: string;
+    children?: string | JSX.Element;
+}
+
+export interface ICartItem {
+    position: TypePosition;
+    amount: number;
+    handleDelete: (position: object) => void;
+}
+
+type TypePosition = {
+    name: string;
+    weigh: number | string;
+    price: number | string;
+};
 export interface ICounter {
     initialValue: number;
     step: number;
@@ -11,4 +29,22 @@ export interface ICounter {
     classNamesText: string;
     classNames: string;
     size: ComponentSize;
+}
+
+export interface IInput {
+    name: string;
+    style: ComponentStyle;
+    classNames: string;
+    classNamesBox: string;
+    onChange: (val: React.ChangeEvent) => void;
+    children: React.ReactNode;
+    type: string;
+    inputName: string;
+}
+
+export interface IStatus {
+    classNames: string | number;
+    trueMessage: string | number;
+    falseMessage: string | number;
+    current: boolean;
 }
